refactored Card Component
