+++ src/components/pages/discovery/Card/index.js
@@ -1,5 +1,3 @@
-import get from 'lodash.get'
-import { requestData } from 'pass-culture-shared'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
@@ -7,22 +5,23 @@ import { withRouter } from 'react-router-dom'
 import withSizes from 'react-sizes'
 import { compose } from 'redux'
 
-import Recto from '../../Recto'
-import Verso from '../../verso'
-import { getHeaderColor } from '../../../utils/colors'
-import currentRecommendationSelector from '../../../selectors/currentRecommendation'
-import nextRecommendationSelector from '../../../selectors/nextRecommendation'
-import previousRecommendationSelector from '../../../selectors/previousRecommendation'
-
-const noop = () => {}
+import { mapDispatchToProps, mapStateToProps } from './connect'
+import Recto from '../../../Recto'
+import Verso from '../../../verso'
+import { getHeaderColor } from '../../../../utils/colors'
 
 export class RawCard extends PureComponent {
+  componentDidMount () {
+    const { handleReadRecommendation, recommendation } = this.props
+    handleReadRecommendation(recommendation)
+  }
+
   componentDidUpdate(prevProps) {
     const {
+      handleClickRecommendation,
       isFlipped,
       recommendation,
-      position,
-      requestDataAction,
+      position
     } = this.props
 
     const isCurrent = recommendation && position === 'current'
@@ -32,21 +31,13 @@ export class RawCard extends PureComponent {
       !prevProps.isFlipped && isFlipped && !recommendation.isClicked
     if (!shouldRequest) return
 
-    const options = {
-      body: { isClicked: true },
-      key: 'recommendations',
-    }
-
-    requestDataAction('PATCH', `recommendations/${recommendation.id}`, options)
+    handleClickRecommendation(recommendation)
   }
 
   render() {
     const { position, recommendation, width } = this.props
-
-    const firstThumbDominantColor = get(
-      recommendation,
-      'firstThumbDominantColor'
-    )
+    const firstThumbDominantColor = recommendation &&
+      recommendation.firstThumbDominantColor
     const headerColor = getHeaderColor(firstThumbDominantColor)
 
     const { index } = recommendation || {}
@@ -73,10 +64,11 @@ RawCard.defaultProps = {
 }
 
 RawCard.propTypes = {
+  handleClickRecommendation: PropTypes.func.isRequired,
+  handleReadRecommendation: PropTypes.func.isRequired,
   isFlipped: PropTypes.bool,
   position: PropTypes.string.isRequired,
   recommendation: PropTypes.object,
-  requestDataAction: PropTypes.func.isRequired,
   width: PropTypes.number.isRequired,
 }
 
@@ -85,37 +77,8 @@ const mapSizeToProps = ({ width, height }) => ({
   width: Math.min(width, 500), // body{max-width: 500px;}
 })
 
-const getSelectorByCardPosition = position => {
-  switch (position) {
-    case 'current':
-      return currentRecommendationSelector
-    case 'previous':
-      return previousRecommendationSelector
-    case 'next':
-      return nextRecommendationSelector
-    default:
-      return noop
-  }
-}
-
 export default compose(
   withSizes(mapSizeToProps),
   withRouter,
-  connect(
-    (state, ownProps) => {
-      const { mediationId, offerId } = ownProps.match.params
-      const recomendationSelector = getSelectorByCardPosition(ownProps.position)
-      const recommendation = recomendationSelector(
-        state,
-        offerId,
-        mediationId,
-        ownProps.position
-      )
-      return {
-        isFlipped: state.verso.isFlipped,
-        recommendation,
-      }
-    },
-    { requestDataAction: requestData }
-  )
+  connect(mapStateToProps, mapDispatchToProps)
 )(RawCard)
