+++ src/components/pages/discovery/Card/tests/index.spec.js
